CIS fix (adding static cast) #jira nojira #rb trivial
[CL 15176925 by Maxwell Hayes in ue5-main branch]

+++ Engine/Plugins/Runtime/Metasound/Source/MetasoundEngine/Private/MetasoundWave.cpp
@@ -202,7 +202,7 @@ namespace Audio
 		IDecoderOutput::FRequirements Reqs
 		{
 			Float32_Interleaved,
-			DecodeBlockSizeInFrames,
+			static_cast<int32>(DecodeBlockSizeInFrames),
 			static_cast<int32>(OutputSampleRate)
 		};
